requirements.txt
aiogram==3.12.0
aiogram_dialog==2.1.0
python-dotenv==1.0.1
